fix: add polyfills adding to repl mode.
