--- libraries/stdlib/js/src/kotlin/js.arrays/sortWith.kt
@@ -1,15 +1,0 @@
-/*
- * Copyright 2010-2021 JetBrains s.r.o. and Kotlin Programming Language contributors.
- * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
- */
-
-package kotlin.js
-
-import kotlin.js.defineSortPolyfillForTypedArrays
-
-@PublishedApi
-@Suppress("NOTHING_TO_INLINE")
-@JsNativeImplementation(defineSortPolyfillForTypedArrays)
-internal inline fun <T> Any.nativeSortWith(noinline comparison: (a: T, b: T) -> Int): Unit {
-    asDynamic().sort(comparison)
-}
